Coverage pumping, fix a bug where the query plan wasn't actually being reused.
